governor little changes
