+++ src/pages/Governor/styles.scss
@@ -111,8 +111,10 @@
           }
 
           & > h4 {
+            @include centered-row;
             @include text-heading6;
             color: var(--color-gray-400);
+            gap: 8px;
             text-transform: uppercase;
 
             @include desktop {
